Refactor msal to use testutils module for tests
