+++ msal/src/test/java/com/microsoft/identity/client/robolectric/shadows/ShadowStrategyResultUnsuccessful.java
@@ -2,7 +2,7 @@ package com.microsoft.identity.client.robolectric.shadows;
 
 import com.microsoft.identity.common.internal.providers.oauth2.TokenResponse;
 import com.microsoft.identity.common.internal.providers.oauth2.TokenResult;
-import com.microsoft.identity.common.internal.providers.ropc.MockTestStrategy;
+import com.microsoft.identity.internal.testutils.strategies.MockTestStrategy;
 
 import org.robolectric.annotation.Implements;
