+++ settings.gradle
@@ -1,7 +1,6 @@
-include ':msal', ':common', ':keyvault', ':labapi'
+include ':msal', ':common', ':testutils'
 project(':common').projectDir = new File('common/common')
-project(':keyvault').projectDir = new File('common/keyvault')
-project(':labapi').projectDir = new File('common/labapi')
+project(':testutils').projectDir = new File('common/testutils')
 
 // test apps
 include ':testapps:sample'
